terraform/lambdasource/requirements.txt
google-auth
requests

